+++ src/main/java/com/gtnewhorizons/angelica/models/template/BlockStaticCube.java
@@ -5,7 +5,7 @@ import com.gtnewhorizons.angelica.api.QuadBuilder;
 import com.gtnewhorizons.angelica.api.QuadProvider;
 import com.gtnewhorizons.angelica.api.QuadView;
 import com.gtnewhorizons.angelica.models.NdQuadBuilder;
-import com.gtnewhorizons.angelica.api.Loader;
+import com.gtnewhorizons.angelica.api.ModelLoader;
 import it.unimi.dsi.fastutil.objects.ObjectImmutableList;
 import me.jellysquid.mods.sodium.client.model.quad.Quad;
 import net.minecraft.block.Block;
@@ -23,7 +23,7 @@ public class BlockStaticCube implements QuadProvider {
 
     public BlockStaticCube(String textureName) {
         this.textureName = textureName;
-        Loader.registerBaker(this::bake);
+        ModelLoader.registerBaker(this::bake);
     }
 
     protected void bake() {
